requirements.txt
fastmcp>=2.6,<4
numpy>=1.26
pydantic-settings>=2.0
redis>=5.0
httpx>=0.27
yfinance>=0.2.40
jugaad-data>=0.33
pytest>=8.0
pytest-asyncio>=0.24
fakeredis>=2.23
beautifulsoup4>=4.12
